Tests/LibWeb/Text/expected/css/HTMLLinkElement-sheet.txt

sheet property initial value: null
Sheet property after stylesheet loaded: [object CSSStyleSheet]
